Split functionality for google sign up and login
