+++ routes/accountVerification.js
@@ -1,39 +1,12 @@
-const { baseURL, nodemailerPass } = require("../config");
+const { baseURL } = require("../config");
 const express = require("express");
 const router = express.Router();
-
-const nodemailer = require("nodemailer");
 const { requireUserAuth } = require("../utils");
+const { sendVerificationEmail } = require("../helpers/account.helper");
 
 router.post("/sendVerificationEmail", requireUserAuth, async (req, res) => {
   try {
-    const { email, userId } = req.body;
-    const transporter = nodemailer.createTransport({
-      service: "gmail",
-      auth: {
-        user: "staynomadhomes@gmail.com",
-        pass: nodemailerPass,
-      },
-    });
-    const userMailOptions = {
-      from: '"NomΛd" <reservations@visitnomad.com>',
-      to: email,
-      subject: `NomΛd email verification needed`,
-      text: `Please click the following link to verify your account
-           ${baseURL}/accountVerification/${email}`,
-      html: `<p>
-            Please click
-            <a href="${baseURL}/accountVerification/${req.body.userId}">here</a>
-            to verify your account
-           </p>`,
-    };
-    transporter.sendMail(userMailOptions, (error, info) => {
-      if (error) {
-        console.log(error);
-      } else {
-        console.log(`Account verification email sent to ${email}`);
-      }
-    });
+    sendVerificationEmail(req.body.email, req.user._id);
     res.status(200).json({
       message: `Verified ${req.user._id}`,
     });
